Move main's save config.
